pyproject.toml
[tool.poetry]
name = "conda-pip-minimal"
version = "0.1.5"
description = "Conda+Pip minimal env exports"
authors = ["Venky Iyer <indigoviolet@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/indigoviolet/conda-pip-minimal"
keywords = ["packaging", "conda", "pip", "environment.yaml"]
classifiers = [
            'Intended Audience :: Developers',
            'Intended Audience :: Science/Research',
            'Topic :: Software Development :: Build Tools'
            ]

[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.6.1"
semver = "^2.13.0"
more-itertools = "^8.14.0"
trio = "^0.22.0"
loguru = "^0.6.0"
importlib-metadata = "^5.0.0"
ruamel-yaml = "^0.17.21"

[tool.poetry.dev-dependencies]
ipython = "^8.5.0"
snoop = "^0.4.2"
trio-typing = "^0.7.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
conda-cpm = 'conda_pip_minimal.cli:app'
